requirements.txt
pandas~=2.2.2
mysql-connector-python
scikit-learn==1.5.0
matplotlib~=3.9.1
seaborn~=0.13.2
keras
tensorflow~=2.17.0
openai~=1.37.1
sqlalchemy
mysqlclient
openai
nltk~=3.8.1
yfinance~=0.2.41
tweepy
sentence-transformers
twarc~=2.14.0
tf-keras
praw~=7.7.1
PyMySQL~=1.1.1
flask~=3.0.3
python-dotenv~=1.0.1
joblib~=1.4.2
numpy~=1.26.4
shap~=0.46.0
xgboost~=2.1.0
imblearn~=0.0
celery
requests~=2.32.3
python-dateutil~=2.9.0.post0
textblob~=0.18.0.post0
Flask-Session
Flask-SocketIO
Flask-Limiter
Flask-SQLAlchemy
Flask-WTF
python-dotenv
redis
cryptography
torch==2.3.1
flask-cors
flask-socketio
dramatiq~=1.17.0
concurrent-log-handler
gunicorn
gevent~=24.2.1
gevent-websocket
pika
greenlet
eventlet~=0.36.1